Split OperationValue into own module
